% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnimalQTLDB.R
\name{AnimalQTLDB}
\alias{AnimalQTLDB}
\title{connect to the local QTL database.
QTL data come from GFF3 format of Animalgenome (https://www.animalgenome.org/cgi-bin/QTLdb/index), in the process of data dump, the nonexistent assignment is NULL, and ultimately build the SQLite database of five data tables,named animalqtldb.db.}
\usage{
AnimalQTLDB()
}
\value{
result
}
\description{
connect to the local QTL database.
QTL data come from GFF3 format of Animalgenome (https://www.animalgenome.org/cgi-bin/QTLdb/index), in the process of data dump, the nonexistent assignment is NULL, and ultimately build the SQLite database of five data tables,named animalqtldb.db.
}
\examples{
AnimalQTLDB()
}
